Mostly working, tests in progress
